Config option: Reduce the number of totemists that spawn in a village.
While this unfortunately can result in a situation where no totemists
will spawn at all, it also mitigates circumstances where you can end up
with a village consisting almost entirely of totemists.

Each structure piece that spawns with a villager spawns a totemist, so
the more village structures (totems & tipii), the more totemists.
Additionally, totemist is a profession that can be randomly chosen.

My solution here was to, on a coinflip, return a totemist, or if that
fails, return the super to generate a random villager.

Additionally, as this is not necessarily something all players of the
mod would want (I just find that villagers on my modded server tend to
be flooded with medicine men), it's currently a configuration option
that defaults to false.